apps/api/requirements.txt
Django>=4.2
djangorestframework
rdkit
torch
numpy<=2
pandas
pillow
dgl
scikit-learn
django-tex
PyJWT
fpdf2
drf-spectacular
networkx==2.8.8
reportlab
cairosvg
python-dotenv
django-cors-headers